backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
email-validator